Added .env.
Added the .env file containing sensitive data. In order to comply with the assignment requirements, it has been included in the repository. However, please note that this is not an optimal security practice and the sensitive data should be stored securely in a separate environment or configuration file in a production setting.